refactor: ECRを作成する部分を切り離し。
